Add specific Symfony command for non-transactional fixtures
Fix tests
