+++ tests/DependencyInjection/Compiler/ConnectionCompilerPassTest.php
@@ -5,6 +5,7 @@ namespace Kununu\TestingBundle\Tests\DependencyInjection\Compiler;
 
 use Kununu\DataFixtures\Executor\ConnectionExecutor;
 use Kununu\DataFixtures\Purger\ConnectionPurger;
+use Kununu\TestingBundle\Command\LoadConnectionFixturesCommand;
 use Kununu\TestingBundle\DependencyInjection\Compiler\AbstractConnectionCompilerPass;
 use Kununu\TestingBundle\DependencyInjection\Compiler\ConnectionCompilerPass;
 
@@ -28,5 +29,10 @@ final class ConnectionCompilerPassTest extends BaseConnectionCompilerPassTestCas
     protected function getExecutorClass(): string
     {
         return ConnectionExecutor::class;
+    }
+
+    protected function getLoadFixturesCommandClass(): string
+    {
+        return LoadConnectionFixturesCommand::class;
     }
 }
